requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pytz==2024.2
canvasapi==3.3.0
requests==2.32.3
twilio==9.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask_bcrypt==1.0.1